test(pytest) Setup Windows
#71
