+++ test/ui-pytest/configs/system.py
@@ -1,0 +1,5 @@
+import platform
+
+IS_LIN = True if platform.uname() == 'Linux' else False
+IS_MAC = True if platform.uname() == 'Darwin' else False
+IS_WIN = True if platform.system() == 'Windows' else False
