+++ test/ui-pytest/configs/timeouts.py
@@ -1,6 +1,7 @@
-# Timoeuts before raising errors
+# Timeouts before raising errors
 
 UI_LOAD_TIMEOUT_SEC = 5
 UI_LOAD_TIMEOUT_MSEC = UI_LOAD_TIMEOUT_SEC * 1000
 PROCESS_TIMEOUT_SEC = 10
+PROCESS_TIMEOUT_MSEC = PROCESS_TIMEOUT_SEC * 1000
 APP_LOAD_TIMEOUT_MSEC = 60
