+++ test/ui-pytest/driver/server.py
@@ -29,15 +29,8 @@ class SquishServer:
         local_system.wait_for_started(_PROCESS_NAME)
 
     @classmethod
-    def stop(cls, attempt: int = 2):
-        local_system.run(['killall', _PROCESS_NAME], check=False)
-        try:
-            local_system.wait_for_close(_PROCESS_NAME, 1)
-        except AssertionError as err:
-            if attempt:
-                cls.stop(attempt-1)
-            else:
-                raise err
+    def stop(cls):
+        local_system.kill_process_by_name(_PROCESS_NAME, verify=False)
 
     # https://doc-snapshots.qt.io/squish/cli-squishserver.html
     def configuring(self, action: str, options: typing.Union[int, str, list]):
