+++ test/ui-pytest/gui/elements/qt/text_label.py
@@ -1,0 +1,8 @@
+from gui.elements.qt.object import QObject
+
+
+class TextLabel(QObject):
+
+    @property
+    def text(self) -> str:
+        return str(self.object.text)
