--- test/ui-pytest/gui/elements/text_label.py
@@ -1,8 +1,0 @@
-from gui.elements.base_object import QObject
-
-
-class TextLabel(QObject):
-
-    @property
-    def text(self) -> str:
-        return str(self.object.text)
