+++ test/ui-pytest/scripts/utils/local_system.py
@@ -1,6 +1,5 @@
 import logging
 import os
-import platform
 import signal
 import subprocess
 import time
@@ -10,6 +9,7 @@ from datetime import datetime
 import psutil
 
 import configs
+from configs.system import IS_WIN
 
 _logger = logging.getLogger(__name__)
 
@@ -20,7 +20,7 @@ def find_process_by_name(process_name: str):
     processes = []
     for proc in psutil.process_iter():
         try:
-            if process_name.lower().split('.')[0] == proc.name().lower():
+            if process_name.lower().split('.')[0] == proc.name().lower().split('.')[0]:
                 processes.append(process_info(
                     proc.pid,
                     proc.name(),
@@ -29,10 +29,10 @@ def find_process_by_name(process_name: str):
         except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
             pass
     if processes:
-        _logger.info(
+        _logger.debug(
             f'Process: {process_name} found in processes list, PID: {", ".join([str(proc.pid) for proc in processes])}')
     else:
-        _logger.info(f'Process: {process_name} not found in processes list')
+        _logger.debug(f'Process: {process_name} not found in processes list')
     return processes
 
 
@@ -41,7 +41,7 @@ def kill_process_by_name(process_name: str, verify: bool = True, timeout_sec: in
     processes = find_process_by_name(process_name)
     for process in processes:
         try:
-            os.kill(process.pid, signal.SIGKILL)
+            os.kill(process.pid, signal.SIGILL if IS_WIN else signal.SIGKILL)
         except PermissionError as err:
             _logger.info(f'Close "{process}" error: {err}')
     if verify and processes:
@@ -90,7 +90,7 @@ def execute(
 
     command = " ".join(str(atr) for atr in command)
     _logger.info(f'Execute: {command}')
-    process = subprocess.Popen(command, shell=shell, stderr=stderr, stdout=stdout)
+    process = subprocess.Popen(command, shell=True, stderr=stderr, stdout=stdout)
     if check and process.returncode != 0:
         stdout, stderr = _get_output(process)
         raise RuntimeError(stderr)
@@ -110,7 +110,3 @@ def run(
     process = subprocess.run(command, shell=shell, stderr=stderr, stdout=stdout, timeout=timeout_sec)
     if check and process.returncode != 0:
         raise subprocess.CalledProcessError(process.returncode, command, process.stdout, process.stderr)
-
-
-def is_mac():
-    return platform.system() == 'Darwin'
