+++ test/ui-pytest/tests/fixtures/aut.py
@@ -30,14 +30,6 @@ def user_data(request) -> system_path.SystemPath:
 
 
 @pytest.fixture
-def user_data_two(request, user_data) -> system_path.SystemPath:
-    user_data = configs.testpath.STATUS_DATA / f'app_{datetime.now():%H%M%S_%f}' / 'data'
-    if hasattr(request, 'param'):
-        system_path.SystemPath(request.param).copy_to(user_data)
-    yield user_data
-
-
-@pytest.fixture
 def main_window(aut: AUT, user_data: system_path.SystemPath):
     aut.launch(f'-d={user_data.parent}')
     yield MainWindow().wait_until_appears().prepare()
